[#37]; refactor: `/my`와 `/member` 분리
